MOD: Refactoring

+++ cegui/include/CEGUI/RendererModules/Direct3D11/RenderTarget.h
@@ -75,13 +75,13 @@ protected:
     //! holds defined area for the RenderTarget
     Rectf d_area;
     //! tangent of the y FOV half-angle; used to calculate viewing distance.
-    static const double d_yfov_tan;
+    static const float d_yfov_tan;
     //! View projection matrix cache
     mutable glm::mat4 d_matrix;
     //! true when d_matrix is valid and up to date
     mutable bool d_matrixValid;
     //! tracks viewing distance (this is set up at the same time as d_matrix)
-    mutable double d_viewDistance;
+    mutable float d_viewDistance;
 };
 
 }
